+++ crates/networking/rpc/eth/transaction.rs
@@ -307,7 +307,8 @@ impl RpcHandler for GetTransactionReceiptRequest {
             block::get_all_block_rpc_receipts(block.header, block.body, storage, Some(index))
                 .await?;
 
-        serde_json::to_value(receipts.last()).map_err(|error| RpcErr::Internal(error.to_string()))
+        serde_json::to_value(receipts.get(index as usize))
+            .map_err(|error| RpcErr::Internal(error.to_string()))
     }
 }